# docs/openapi/v1.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Betsi Patient Flow API",
    "description": "Patient flow and escalation for emergency departments. Errors are RFC 9457 problem details with a stable \u0027code\u0027. See docs/API.md and docs/API-VERSIONING.md.",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/episodes/{episodeId}": {
      "get": {
        "tags": [
          "Episodes"
        ],
        "summary": "One episode: demographics, state, timings and its escalations.",
        "parameters": [
          {
            "name": "episodeId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/boards/waiting": {
      "get": {
        "tags": [
          "Episodes"
        ],
        "summary": "Patients waiting or awaiting treatment, longest wait first, a page at a time.",
        "parameters": [
          {
            "name": "locationId",
            "in": "query",
            "description": "Only patients at this location.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "\u0060Waiting\u0060 or \u0060AwaitingTreatment\u0060. Both if omitted.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "minWaitingMinutes",
            "in": "query",
            "description": "Only patients who have waited at least this long.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "minAgeYears",
            "in": "query",
            "description": "Only patients at least this old.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "maxAgeYears",
            "in": "query",
            "description": "Only patients at most this old, e.g. 17 for paediatrics.",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The \u0060nextCursor\u0060 from the previous page.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "description": "1\u2013200, default 50.",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WaitingBoardPage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EpisodeDetail": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "ageYears": {
            "type": "integer",
            "format": "int32"
          },
          "nhsNumber": {
            "type": "string",
            "nullable": true
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "locationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "arrivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "triageStartedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "treatmentStartedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "minutesWaiting": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "escalations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeEscalationSummary"
            },
            "nullable": true
          },
          "openFollowUpExceptions": {
            "type": "integer",
            "format": "int32"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "The canonical view of one episode (spec \u00A74.1: \u0060GET /episodes/{id}\u0060)."
      },
      "EpisodeEscalationSummary": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "trigger": {
            "type": "string",
            "nullable": true
          },
          "responsibleRole": {
            "type": "string",
            "nullable": true
          },
          "tierLevel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "acknowledgementDueAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "acknowledgementOverdue": {
            "type": "boolean"
          },
          "acknowledgedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "resolvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": {}
      },
...
      "WaitingBoardPage": {
        "type": "object",
        "properties": {
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WaitingBoardRow"
            },
            "nullable": true
          },
          "nextCursor": {
            "type": "string",
            "description": "Pass as \u0060cursor\u0060 for the next page; null on the last page.",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WaitingBoardRow": {
        "type": "object",
        "properties": {
          "episodeId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "ageYears": {
            "type": "integer",
            "format": "int32"
          },
          "state": {
            "type": "string",
            "nullable": true
          },
          "locationId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "arrivedAt": {
            "type": "string",
            "format": "date-time"
          },
          "minutesWaiting": {
            "type": "integer",
            "format": "int32"
          },
          "openEscalations": {
            "type": "integer",
            "format": "int32"
          },
          "highestTierRaised": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "acknowledgementOverdue": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "X-Betsi-Tenant": {
        "type": "apiKey",
        "description": "Tenant ID. Development tenants: 11111111-1111-1111-1111-111111111111 or 22222222-2222-2222-2222-222222222222.",
        "name": "X-Betsi-Tenant",
        "in": "header"
      },
      "X-Betsi-Actor-Role": {
        "type": "apiKey",
        "description": "Actor role, e.g. Nurse.",
        "name": "X-Betsi-Actor-Role",
        "in": "header"
      },
      "X-Betsi-Actor": {
        "type": "apiKey",
        "description": "Optional actor ID (GUID).",
        "name": "X-Betsi-Actor",
        "in": "header"
      },
      "Bearer": {
        "type": "http",
        "description": "An access token from the configured OIDC provider. In Development the X-Betsi headers may be used instead.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [],
      "X-Betsi-Tenant": [],
      "X-Betsi-Actor-Role": [],
      "X-Betsi-Actor": []
    }
  ],
...
}